+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a case where `sat xname2nid` and `sat nid2xname` failed when the input had
   extraneous whitespace.
 - Fixed `--logfile` to accept a file name without leading directories.
+- Fixed issues with filter error messages being printed twice or not being printed.
 
 ### Added
 - Added a ``--fields`` option to allow displaying only specific fields